run offline

+++ src/AutorDAO.java
@@ -8,9 +8,10 @@ import java.util.List;
 // Classe para acesso ao banco de dados
 public class AutorDAO {
     private Connection connection;
-
+    private boolean offline;
     // Construtor: conecta ao banco
     public AutorDAO() {
+        offline = true;
         try {
             // Carrega driver JDBC
             Class.forName("com.mysql.cj.jdbc.Driver");
@@ -21,10 +22,12 @@ public class AutorDAO {
 
         try {
             // Conecta ao banco
-            connection = DriverManager.getConnection(
-                    "jdbc:mysql://localhost:3307/biblioteca",
-                    "root",
-                    "senha");
+            if (!offline) {
+                connection = DriverManager.getConnection(
+                        "jdbc:mysql://localhost:3307/biblioteca",
+                        "root",
+                        "senha");
+            }
         } catch (SQLException e) {
             System.out.println("Erro");
             e.printStackTrace();
@@ -45,6 +48,7 @@ public class AutorDAO {
 
     // Insere autor
     public void inserirAutor(Autor autor) {
+        if(offline) return;
         String sql = "INSERT INTO autores (nome, informacoes) VALUES (?, ?)";
         try (PreparedStatement stmt = connection.prepareStatement(sql)) {
             stmt.setString(1, autor.getNome());
@@ -57,7 +61,10 @@ public class AutorDAO {
 
     // Lista autores ativos
     public List<Autor> listarAutores() {
+
         List<Autor> lista = new ArrayList<>();
+
+        if (offline)return lista;
         String sql = "SELECT * FROM autores WHERE excluido = FALSE";
         try (Statement stmt = connection.createStatement();
                 ResultSet rs = stmt.executeQuery(sql)) {
@@ -77,6 +84,7 @@ public class AutorDAO {
 
     // Atualiza autor pelo ID
     public void atualizarAutor(Autor autor) {
+        if (offline) return;
         String sql = "UPDATE autores SET nome = ?, informacoes = ? WHERE id = ?";
         try (PreparedStatement stmt = connection.prepareStatement(sql)) {
             stmt.setString(1, autor.getNome());
@@ -90,6 +98,7 @@ public class AutorDAO {
 
     // Exclui o autor
     public void excluirAutor(int id) {
+        if (offline) return;
         String sql = "DELETE FROM autores WHERE id = ?;";
         try (PreparedStatement stmt = connection.prepareStatement(sql)) {
             stmt.setInt(1, id);
